Igor v2

+++ src/main/java/com/example/accountingsystem/entities/contract/ContractService.java
@@ -5,12 +5,14 @@ import com.example.accountingsystem.entities.user.User;
 import org.apache.commons.beanutils.BeanUtils;
 import org.mapstruct.factory.Mappers;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import java.lang.reflect.InvocationTargetException;
 import java.sql.Date;
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -28,29 +30,39 @@ public class ContractService {
         this.mapper = Mappers.getMapper(ContractMapper.class);
     }
 
-    public List<ContractDTO> getContractsForUser(long id) {
-        User user = userDetailsService.getUserById(id);
+    public List<ContractDTO> getContractsForUser() {
+        User user = userDetailsService.getCurrentUser();
         if (user == null) {
             return null;
         }
         if (user.getRole() == User.Role.ADMIN) {
             return mapper.toListOfDTO(contractRepo.findAll());
         }
-        List<Contract> entities = contractRepo.getContractByAssociatedUserId(id);
+        List<Contract> entities = contractRepo.getContractByAssociatedUserId(user.getId());
         return mapper.toListOfDTO(entities);
     }
 
-    public void addContract(ContractDTO dto) {
+    public boolean addContract(ContractDTO dto) {
+        User currentUser = userDetailsService.getCurrentUser();
+        if (dto.userId != currentUser.getId() && currentUser.getRole() != User.Role.ADMIN) {
+            return false;
+        }
         Contract contract = mapper.DTOtoContract(dto);
         contract.setAssociatedUser(userDetailsService.getCurrentUser());
+
         contractRepo.save(contract);
+        return true;
     }
 
     public ContractDTO getContractDtoById(long id) {
         Optional<Contract> opt = contractRepo.findById(id);
         ContractDTO dto = null;
         if (opt.isPresent()) {
-            dto = mapper.contractContractToDTO(opt.get());
+            Contract contract = opt.get();
+            if (!Objects.equals(contract.getAssociatedUser().getId(), userDetailsService.getCurrentUser().getId())) {
+                return null;
+            }
+            dto = mapper.contractContractToDTO(contract);
         }
         return dto;
     }
@@ -61,15 +73,30 @@ public class ContractService {
     }
 
     public List<ContractDTO> getContractsByGivenPeriod(LocalDate beginDate, LocalDate endDate) {
-        List<Contract> entities = contractRepo.getContractsByGivenPeriod(Date.valueOf(beginDate), Date.valueOf(endDate));
+        User currentUser = userDetailsService.getCurrentUser();
+        List<Contract> entities;
+        if (currentUser.getRole() == User.Role.ADMIN) {
+            entities = contractRepo.getContractsByGivenPeriod(Date.valueOf(beginDate), Date.valueOf(endDate));
+        }
+        else {
+            entities = contractRepo.getContractsByGivenPeriod(Date.valueOf(beginDate), Date.valueOf(endDate), currentUser.getId());
+        }
         return mapper.toListOfDTO(entities);
     }
 
-    public void updateContract(ContractDTO dto) {
+    public boolean updateContract(ContractDTO dto) {
+        User currentUser = userDetailsService.getCurrentUser();
         long id = dto.id;
         Contract updatingContract = mapper.DTOtoContract(dto);
         updatingContract.setAssociatedUser(userDetailsService.getUserById(dto.userId));
         Contract contractToBeUpdated = getContractById(id);
+
+        if (currentUser.getRole() != User.Role.ADMIN) {
+            if (dto.userId != contractToBeUpdated.getAssociatedUser().getId() || dto.userId != currentUser.getId()) {
+                return false;
+            }
+        }
+
         if (contractToBeUpdated != null) {
             try {
                 BeanUtils.copyProperties(contractToBeUpdated, updatingContract);
@@ -78,13 +105,25 @@ public class ContractService {
             }
             contractToBeUpdated.setId(id);
             contractRepo.save(contractToBeUpdated);
+            return true;
         }
         else {
-            contractRepo.save(updatingContract);
+            return false;
         }
     }
 
-    public void deleteContract(long id) {
-        contractRepo.deleteById(id);
+    public boolean deleteContract(long id) {
+        Optional<Contract> opt = contractRepo.findById(id);
+        User currentUser = userDetailsService.getCurrentUser();
+        if (opt.isPresent()) {
+            if (currentUser.getRole() != User.Role.ADMIN) {
+                if (!Objects.equals(opt.get().getAssociatedUser().getId(), currentUser.getId())) {
+                    return false;
+                }
+            }
+            contractRepo.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
